requirements.txt
asgiref==3.7.2
bleach==6.0.0
cffi==1.15.1
cryptography==41.0.0
Django==4.2.1
django-encrypted-model-fields==0.6.5
django-environ==0.10.0
Markdown==3.4.3
pycparser==2.21
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
webencodings==0.5.1
